# Conflicts:
#	resources/js/Composables/getIrisComponents.ts
#	routes/iris/root.php

+++ app/Actions/Catalogue/ProductCategory/UI/IndexFamilies.php
@@ -34,6 +34,9 @@ use Inertia\Inertia;
 use Inertia\Response;
 use Lorisleiva\Actions\ActionRequest;
 use Spatie\QueryBuilder\AllowedFilter;
+use Spatie\QueryBuilder\AllowedSort;
+use Spatie\QueryBuilder\Sorts\Sort;
+use Illuminate\Database\Eloquent\Builder;
 
 class IndexFamilies extends OrgAction
 {
@@ -112,7 +115,6 @@ class IndexFamilies extends OrgAction
             );
         }
 
-
         $queryBuilder->leftJoin('shops', 'product_categories.shop_id', 'shops.id');
         $queryBuilder->leftJoin('organisations', 'product_categories.organisation_id', '=', 'organisations.id');
         $queryBuilder->leftJoin('product_category_sales_intervals', 'product_category_sales_intervals.product_category_id', 'product_categories.id');
@@ -129,11 +131,9 @@ class IndexFamilies extends OrgAction
             } elseif ($parent->type == ProductCategoryTypeEnum::SUB_DEPARTMENT) {
                 $queryBuilder->where('product_categories.sub_department_id', $parent->id);
             } else {
-                // todo
                 abort(419);
             }
         }
-
 
         return $queryBuilder
             ->defaultSort('product_categories.code')
@@ -161,7 +161,9 @@ class IndexFamilies extends OrgAction
                 'organisations.slug as organisation_slug',
                 'product_category_sales_intervals.sales_grp_currency_all as sales_all',
                 'product_category_ordering_intervals.invoices_all as invoices_all',
-                DB::raw("(
+                'product_categories.master_product_category_id',
+                DB::raw(
+                    "(
                     SELECT json_agg(json_build_object(
                         'id', c.id,
                         'slug', c.slug,
@@ -173,13 +175,44 @@ class IndexFamilies extends OrgAction
                     WHERE chm.model_id = product_categories.id
                         AND chm.model_type = 'ProductCategory'
                         AND c.deleted_at IS NULL
-                ) as collections"),
+                )::text as collections"
+                ),
             ])
             ->leftJoin('product_category_stats', 'product_categories.id', 'product_category_stats.product_category_id')
             ->where('product_categories.type', ProductCategoryTypeEnum::FAMILY)
             ->leftjoin('product_categories as departments', 'departments.id', 'product_categories.department_id')
             ->leftjoin('product_categories as sub_departments', 'sub_departments.id', 'product_categories.sub_department_id')
-            ->allowedSorts(['code', 'name', 'shop_code', 'department_code', 'number_current_products', 'sub_department_name', 'department_name'])
+            ->allowedSorts([
+                'code',
+                'name',
+                'shop_code',
+                'department_code',
+                'number_current_products',
+                'sub_department_name',
+                'department_name',
+                AllowedSort::custom(
+                    'collections',
+                    new class() implements Sort {
+                        public function __invoke(Builder $query, bool $descending, string $property)
+                        {
+                            $direction = $descending ? 'desc' : 'asc';
+                            $query->orderBy(
+                                DB::raw(
+                                    "(
+                                SELECT json_agg(c.name)
+                                FROM collection_has_models chm
+                                JOIN collections c ON chm.collection_id = c.id
+                                WHERE chm.model_id = product_categories.id
+                                AND chm.model_type = 'ProductCategory'
+                                AND c.deleted_at IS NULL
+                            )::text"
+                                ),
+                                $direction
+                            );
+                        }
+                    }
+                )
+            ])
             ->allowedFilters([$globalSearch])
             ->withPaginator($prefix, tableName: request()->route()->getName())
             ->withQueryString();
@@ -281,25 +314,23 @@ class IndexFamilies extends OrgAction
     public function getActions(ActionRequest $request): array
     {
         $actions = [];
-        if ($this->canEdit) {
-            if ($this->parent instanceof ProductCategory) {
-                $createRoute = "grp.org.shops.show.catalogue.departments.show.families.create";
+        if ($this->canEdit && $this->parent instanceof ProductCategory) {
+            $createRoute = "grp.org.shops.show.catalogue.departments.show.families.create";
 
-                if ($this->parent->type == ProductCategoryTypeEnum::SUB_DEPARTMENT) {
-                    $createRoute = "grp.org.shops.show.catalogue.departments.show.sub_departments.show.family.create";
-                }
-
-                $actions[] = [
-                    'type'    => 'button',
-                    'style'   => 'create',
-                    'tooltip' => __('new family'),
-                    'label'   => __('family'),
-                    'route'   => [
-                        'name'       => $createRoute,
-                        'parameters' => $request->route()->originalParameters()
-                    ]
-                ];
+            if ($this->parent->type == ProductCategoryTypeEnum::SUB_DEPARTMENT) {
+                $createRoute = "grp.org.shops.show.catalogue.departments.show.sub_departments.show.family.create";
             }
+
+            $actions[] = [
+                'type'    => 'button',
+                'style'   => 'create',
+                'tooltip' => __('new family'),
+                'label'   => __('family'),
+                'route'   => [
+                    'name'       => $createRoute,
+                    'parameters' => $request->route()->originalParameters()
+                ]
+            ];
         }
 
 
@@ -342,9 +373,6 @@ class IndexFamilies extends OrgAction
                     'icon'  => ['fal', 'fa-folder-tree'],
                     'title' => __('department')
                 ];
-                // $iconRight  = [
-                //     'icon' => 'fal fa-folder',
-                // ];
                 $iconRight  = $this->parent->state->stateIcon()[$this->parent->state->value];
                 $afterTitle = [
 
@@ -357,9 +385,6 @@ class IndexFamilies extends OrgAction
                     'icon'  => ['fal', 'fa-dot-circle'],
                     'title' => __('sub department')
                 ];
-                // $iconRight  = [
-                //     'icon' => 'fal fa-folder',
-                // ];
                 $iconRight  = $this->parent->state->stateIcon()[$this->parent->state->value];
                 $afterTitle = [
 
@@ -370,30 +395,28 @@ class IndexFamilies extends OrgAction
 
         $routes = null;
 
-        if ($this->parent instanceof PRoductCategory) {
-            if ($this->parent->type == ProductCategoryTypeEnum::SUB_DEPARTMENT) {
-                $routes = [
-                    'attach' => [
-                        'name'       => 'grp.models.sub-department.families.attach',
-                        'parameters' => [
-                            'subDepartment' => $this->parent->id
-                        ]
-                    ],
-                    'detach' => [
-                        'method'  => 'delete',
-                        'name'       => 'grp.models.sub-department.family.detach',
-                        'parameters' => [
-                            'subDepartment' => $this->parent->id
-                        ]
-                    ],
-                    'fetch_families'   => [
-                        'name'      =>  'grp.json.product_category.families.index',
-                        'parameters' => [
-                            'productCategory' => $this->parent->slug
-                        ]
-                    ],
-                ];
-            }
+        if ($this->parent instanceof ProductCategory && $this->parent->type == ProductCategoryTypeEnum::SUB_DEPARTMENT) {
+            $routes = [
+                'attach'         => [
+                    'name'       => 'grp.models.sub-department.families.attach',
+                    'parameters' => [
+                        'subDepartment' => $this->parent->id
+                    ]
+                ],
+                'detach'         => [
+                    'method'     => 'delete',
+                    'name'       => 'grp.models.sub-department.family.detach',
+                    'parameters' => [
+                        'subDepartment' => $this->parent->id
+                    ]
+                ],
+                'fetch_families' => [
+                    'name'       => 'grp.json.product_category.families.index',
+                    'parameters' => [
+                        'productCategory' => $this->parent->slug
+                    ]
+                ],
+            ];
         }
 
 
@@ -422,12 +445,12 @@ class IndexFamilies extends OrgAction
                     'navigation' => $navigation,
                 ],
                 ProductCategoryTabsEnum::INDEX->value => $this->tab == ProductCategoryTabsEnum::INDEX->value ?
-                    fn () => FamiliesResource::collection($families)
-                    : Inertia::lazy(fn () => FamiliesResource::collection($families)),
+                    fn() => FamiliesResource::collection($families)
+                    : Inertia::lazy(fn() => FamiliesResource::collection($families)),
 
                 ProductCategoryTabsEnum::SALES->value => $this->tab == ProductCategoryTabsEnum::SALES->value ?
-                    fn () => FamiliesResource::collection($families)
-                    : Inertia::lazy(fn () => FamiliesResource::collection($families)),
+                    fn() => FamiliesResource::collection($families)
+                    : Inertia::lazy(fn() => FamiliesResource::collection($families)),
             ]
         )->table($this->tableStructure(parent: $this->parent, modelOperations: null, canEdit: false, prefix: ProductCategoryTabsEnum::INDEX->value, sales: false))
             ->table($this->tableStructure(parent: $this->parent, modelOperations: null, canEdit: false, prefix: ProductCategoryTabsEnum::SALES->value, sales: $this->sales));
